pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "credit-risk-calculator"
version = "0.1.0"
description = "Credit Risk Scoring API and ML pipeline."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
]

dependencies = [
    "pandas==1.5.3",
    "scikit-learn==1.3.2",
    "xgboost==1.7.6",
    "mlflow==2.9.2",
    "matplotlib==3.7.2",
    "seaborn==0.13.0",
    "joblib==1.3.2",
    "fastapi==0.110.0",
    "uvicorn[standard]==0.25.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "gunicorn==21.2.0",
    "requests==2.31.0",
    "optuna==3.5.0",
    "prometheus-fastapi-instrumentator==6.1.0",
    "opentelemetry-api==1.22.0",
    "opentelemetry-sdk==1.22.0",
    "opentelemetry-exporter-otlp==1.22.0",
    "opentelemetry-instrumentation-fastapi==0.44b0",
    "opentelemetry-instrumentation-logging==0.44b0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "mypy==1.8.0",
    "ruff==0.3.4",
    "jupyter==1.0.0",
    "pip-tools==7.4.1",
    "mkdocs-material==9.5.10",
]

[tool.setuptools]
packages = {find = {where = ["src"]}}

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true 